WIP move towards hotcode reloading
