+++ src/test.h
@@ -8,7 +8,7 @@
 #include <vector>
 #include <functional>
 
-#include "main.h"
+#include "game.h"
 
 #define PP_CAT(a, b) PP_CAT_I(a, b)
 #define PP_CAT_I(a, b) PP_CAT_II(a ## b)
